fix: preserve typed not-found for blank session selectors
